Decided (probably) on program arragement for new setup
