+++ check.py
@@ -1,0 +1,13 @@
+import sqlite3
+import time
+
+'''
+Price-Search Copyright (C) 2021 Alex Henderson
+Version Alpha 2.0
+
+CHECK file holds functions to actively search through
+live price database and record descrepancies found to
+a seperate database. 
+
+'''
+
